+++ packages/coin-cro/package.json
@@ -1,13 +1,13 @@
 {
-  "name": "@coolwallet/cro",
+  "name": "@coolwallet/cronos",
   "version": "1.0.0",
-  "description": "Coolwallet Crypto.org Chain sdk",
+  "description": "Coolwallet Cronos sdk",
   "main": "lib/index.js",
   "author": "coolwallet-team",
   "license": "Apache-2.0",
   "scripts": {
     "build:types": "tsc --emitDeclarationOnly",
-    "build:ts": "babel src --out-dir lib --extensions \".ts,.tsx,.js\" --source-maps inline",
+    "build:ts": "babel src --out-dir lib --extensions \".ts,.tsx\" --source-maps inline",
     "build": "npm run build:types && npm run build:ts",
     "test": "jest --verbose"
   },
@@ -30,18 +30,11 @@
     "url": "https://github.com/CoolBitX-Technology/coolwallet3-sdk/issues"
   },
   "dependencies": {
-    "@cosmostation/cosmosjs": "^0.11.1",
     "@types/keccak": "^3.0.1",
-    "bech32": "^2.0.0",
     "bip32": "^2.0.4",
     "eth-sig-util": "^3.0.1",
     "keccak": "^3.0.2",
-    "protobufjs": "^6.11.2",
-    "protocol-buffers": "^4.2.0",
-    "protocol-buffers-encodings": "^1.1.1",
     "rlp": "^2.2.7",
-    "varint": "^6.0.0",
-    "varstruct": "^6.1.3",
     "web3": "^1.7.0"
   }
 }
