+++ packages/coin-cro/src/config/params.ts
@@ -1,11 +1,27 @@
-export const COIN_TYPE = '8000018a';
+export const COIN_TYPE = '8000003c';
+
+export const CHAIN_ID = 25;
 
 export const TRANSFER = {
-    script: '03030201C707000000018ACC07100aA00700CC07100aA00700CC07C01E0a1c2f636f736d6f732e62616e6b2e763162657461312e4d736753656e64CC071012A00700CC07100aBFACD7C021FFC002CC071012BFACD7C061FFC002CC07101aA00700CC07C0090a076261736563726fCC071012A00700BAACC76CA1080D01BE0700BE0700BE0700BE0700CC071012A0070029AC9700C9CAAC9700C9BE0700BE0700CC071012A00700CC07100aA00700CC07C0270a460a1f2f636f736d6f732e63727970746f2e736563703235366b312e5075624b657912230a21CAA0C70021CC07C00612040a020801CC071018BFACC700C108BE0700CC071012A00700CC07100aA00700CC07C0090a076261736563726fCC071012A00700BAACC76CA9080D01BE0700BE0700CC071010BFACC700B108BE0700BE0700CC07C01C1a1A63727970746f2d6f72672d636861696e2d6d61696e6e65742d31CC071020BFACC700B908DC07C00343524FDDACD70061FFC0DAACC7C0A10808D207CC05065052455353425554546F4E',
-    signature: '304502206FD556F37F2390835CA56A91CCB9E7005FC56C74551FEB8CB270BF835C1B7B1B0221008838C3BE0E2237ED3C7226418D08802501337A1078142E11D4FD8198BDF2B5C8'
-}
+  script: `03040601C707000000003CA00700C2ACD70032FFF8C2ACD7001EFFF6C2ACD70028FFF6CC071094CAA02700C2A2D700FFF6CC071080CC0E1019C2E09700CC07C0028080BE0710DC07C00343524FCC0FC0023078BAA02F6C0E04DDF09700DAA2D7C0FFF612D207CC05065052455353425554546F4E`,
+  signature: `003045022100B3D3C5F089BFD31021357B05F769256299460DAA54B4DBD8777158C047A3729E02206078F1403E17FB750E7B4BAE197566165AC5F5061C0764F325AC7327EC2B5EB9`,
+  get scriptWithSignature(): string {
+    return this.script + this.signature;
+  },
+};
 
-export enum TX_TYPE_URL{
-    SEND = '/cosmos.bank.v1beta1.MsgSend',
-}
+export const ERC20 = {
+  script: `03040601C707000000003CA00700C2ACD70034FFF8C2A5D700FFF6C2ACD7002AFFF6CC071094CAAC270045CC07C01380B844a9059cbb000000000000000000000000CAA02700CC07200000000000000000000000000000000000000000CAA2C7000CCC0E1019C2E09700CC07C0028080BE0710DC07C00343524F11ACC7CC3C1D04591507C004CC0F104012AC17C03D0401071507C002FF00B5AC17003DCAACBF003EDEF09700250F00CC0FC0023078BAA02F6C0E04DDF0970012AC17C03C0400141507C002FF00B5AC17003CDAA2C7B00CD207CC05065052455353425554546F4E`,
+  signature: `00304502207254B9195694D839E5A385F7B787CD316C03738006A8F4E0978FD351AADF20C9022100DAE520982AABC5B0D6C03F9D1BA4A99B17B9D05E17ECEEDFC6CC7E8E3786EF04`,
+  get scriptWithSignature(): string {
+    return this.script + this.signature;
+  },
+};
 
+export const SmartContract = {
+  script: '',
+  signature: ''.padStart(144, '0'),
+  get scriptWithSignature(): string {
+    return this.script + this.signature;
+  },
+};
