+++ packages/coin-cro/src/config/tokenType.ts
