+++ packages/coin-cro/src/utils/scriptUtils.ts
