+++ packages/coin-cro/src/utils/stringUtil.ts
