+++ packages/coin-cro/webpack.config.js
