+++ packages/coin-org/src/config/params.ts
@@ -1,0 +1,11 @@
+export const COIN_TYPE = '8000018a';
+
+export const TRANSFER = {
+    script: '03030201C707000000018ACC07100aA00700CC07100aA00700CC07C01E0a1c2f636f736d6f732e62616e6b2e763162657461312e4d736753656e64CC071012A00700CC07100aBFACD7C021FFC002CC071012BFACD7C061FFC002CC07101aA00700CC07C0090a076261736563726fCC071012A00700BAACC76CA1080D01BE0700BE0700BE0700BE0700CC071012A0070029AC9700C9CAAC9700C9BE0700BE0700CC071012A00700CC07100aA00700CC07C0270a460a1f2f636f736d6f732e63727970746f2e736563703235366b312e5075624b657912230a21CAA0C70021CC07C00612040a020801CC071018BFACC700C108BE0700CC071012A00700CC07100aA00700CC07C0090a076261736563726fCC071012A00700BAACC76CA9080D01BE0700BE0700CC071010BFACC700B108BE0700BE0700CC07C01C1a1A63727970746f2d6f72672d636861696e2d6d61696e6e65742d31CC071020BFACC700B908DC07C00343524FDDACD70061FFC0DAACC7C0A10808D207CC05065052455353425554546F4E',
+    signature: '304502206FD556F37F2390835CA56A91CCB9E7005FC56C74551FEB8CB270BF835C1B7B1B0221008838C3BE0E2237ED3C7226418D08802501337A1078142E11D4FD8198BDF2B5C8'
+}
+
+export enum TX_TYPE_URL{
+    SEND = '/cosmos.bank.v1beta1.MsgSend',
+}
+
