+++ packages/coin-org/src/utils/scriptUtil.ts
