+++ packages/coin-tester/src/components/coins/coin-cronos/index.tsx
@@ -4,8 +4,8 @@ import { Transport, apdu, utils, config } from '@coolwallet/core';
 import { NoInput, OneInput, TwoInputs } from '../../../utils/componentMaker';
 import Web3 from 'web3';
 
-import Cronos from '@coolwallet/coin-cronos';
-import { Transaction } from '@coolwallet/coin-cronos/lib/config/types';
+import Cronos from '@coolwallet/coin-cro';
+import { Transaction } from '@coolwallet/coin-cro/lib/config/types';
 
 const web3 = new Web3('https://evm-cronos.crypto.org');
